Refactored create and search recipe so that delete recipe is a function of search recipe and only appears when embedded on create recipe page
