+++ src/app/create-recipe/create-recipe.component.ts
@@ -232,61 +232,6 @@ export class CreateRecipeComponent {
         }
 
         /**
-         * Delete a document in a specific location
-         * @param {any} query
-         */
-        public deleteDoc(query) {
-          this.dbHelp.deleteDocWhere('users/'+this.userInfo+'/recipeList/', query);
-        }
-
-        /**
-         * splice the temp at the index
-         * @param {any} temp
-         * @param {any} index
-         */
-        public tempSplice(temp, index) {
-          temp.splice(index, 1);
-        }
-
-        /**
-         * @param {any} temp
-         */
-        public setLocalStorageDelete(temp) {
-          localStorage.setItem('cachedRecipes', JSON.stringify(temp));
-          this.search.fetchCache();
-        }
-
-        /**
-         * @param {string} message
-         * @return {any}
-         */
-        public askConfirm(message) {
-          return confirm(message);
-        }
-
-        /**
-         * A funtion that remove a recipe from the databse
-         * @param {any} recipe Recipe being deleted
-         * @param {any} r Holds boolean value of the confirmation popup
-         */
-        public deleteRecipe(recipe) {
-          const r = this.askConfirm('Are you sure you want to delete this recipe?');
-          if (r == true) {
-            console.log(recipe);
-            const query = 'recipeName:==:'+ recipe.recipeName+'';
-
-            this.deleteDoc(query);
-
-            localStorage.setItem('updatePending', 'true');
-            const temp: Array<any> = JSON.parse(localStorage.getItem('cachedRecipes'));
-            const index = temp.findIndex((index) => index.recipeName === recipe.recipeName);
-
-            this.tempSplice(temp, index);
-            this.setLocalStorageDelete(temp);
-          }
-        }
-
-        /**
          * Test function that prints fuzzyResults to console
          */
         public logResults() {
